refactor: extract the shared hint roundtrip test helpers
wait_for_hint_roundtrip / format_peer_snapshot が fake.rs と iroh/tests/mod.rs に
逐語重複していたため、#[cfg(test)] の crate 内共有モジュール test_support へ集約する。
fake 側は expected_source_peer: None で従来と同一挙動(WP-B15)。
